Splitted mappings file for tagging
